backend/requirements.txt
Django==4.2.4
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
psycopg==3.1.14
psycopg2-binary==2.9.9
python-dotenv==0.19.2
requests==2.31.0